Remove delegated event listeners
When I changed "presenter-toggle" to "presenter" I messed this up. The
toggle is inside the other and has no children so it always used to
work. But with the change, I should have been checking to see whether
the target or any of its ancestors matched the id. That's too much work,
so I am switching to dumb individual event listeners.